requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
langchain==0.3.7
langchain-openai==0.2.9
langchain-chroma==0.1.4
langchain-community==0.3.7
langchain-text-splitters==0.3.2
unstructured==0.16.10
pypdf==5.1.0
pydantic-settings==2.6.1
